move maybe_match_name from com to ops
