+++ pandas/core/common.py
@@ -121,21 +121,6 @@ def _consensus_name_attr(objs):
     return name
 
 
-def _maybe_match_name(a, b):
-    a_has = hasattr(a, 'name')
-    b_has = hasattr(b, 'name')
-    if a_has and b_has:
-        if a.name == b.name:
-            return a.name
-        else:
-            return None
-    elif a_has:
-        return a.name
-    elif b_has:
-        return b.name
-    return None
-
-
 def _get_info_slice(obj, indexer):
     """Slice the info axis of `obj` with `indexer`."""
     if not hasattr(obj, '_info_axis_number'):
